==> Lessons/03_DataExploration_Part2.Rmd <==
---
title: "3: Data Exploration"
author: "Environmental Data Analytics | John Fay & Luana Lima"
date: "Fall 2024"
output: pdf_document
editor_options: 
  chunk_output_type: console
---

## Lesson Objectives
1. Import and explore datasets in R
2. Graphically explore datasets in R
3. Apply data exploration skills to a real-world example dataset

---

## Discussion: Why do we explore our data?

Why is data exploration our first step in analyzing a dataset? What information do we gain? How does data exploration aid in our decision-making for data analysis steps further down the pipeline?

>Answer: 

---

## Import data and view summaries

```{r, Setup-Import-Explore, message = FALSE}
## -Set up-
# 1. Load packages (tidyverse, here)
library(tidyverse)
library(here)

# 2. Check the working directory
getwd()
here()

# 3. Import datasets
USGS.flow.data <- read.csv(
  file = here("Data/Processed/USGS_Site02085000_Flow_Processed.csv"),
  stringsAsFactors = TRUE,
  colClasses = c('site_no'='factor','datetime'='factor')
  )

## -Initial Exploration-

#View(USGS.flow.data)
# Alternate option: click on data frame in Environment tab

#View column names and the structure of the dataset
colnames(USGS.flow.data)
str(USGS.flow.data)

# Check our date column: is it a date object?
class(USGS.flow.data$datetime)

# Change it to a date object and confirm
USGS.flow.data$datetime <- as.Date(USGS.flow.data$datetime, format='%Y-%m-%d')

```

---

## Visualization for Data Exploration
Although the `summary()` function is helpful in getting an idea of the spread of values in a numeric dataset, it can be useful to create visual representations of the data to help form hypotheses and direct downstream data analysis. Below is a summary of the useful types of graphs for data exploration.

Note: each of these approaches utilize the package "ggplot2". We will be covering the syntax of ggplot in a later lesson, but for now you should familiarize yourself with the functionality of what each command is doing.

### Bar Chart (function: geom_bar)

Visualize count data for categorical variables. 

```{r, fig.height = 3, fig.width = 4}
ggplot(data = USGS.flow.data, aes(x = discharge.mean.approval)) +
  geom_bar()
```

### Histogram (function: geom_histogram)
Visualize distributions of values for continuous numerical variables. What is happening in each line of code? Insert a comment above each line.

```{r, fig.height = 3, fig.width = 4}
# plot a histogram
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = discharge.mean))

# change width of bins (bars)
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = discharge.mean), binwidth = 10)

# change number of bins (bars)
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = discharge.mean), bins = 20)

# change range of values in x axis
ggplot(USGS.flow.data, aes(x = discharge.mean)) +
  geom_histogram(binwidth = 10) + 
  scale_x_continuous(limits = c(0, 500))
  
# different x value
ggplot(USGS.flow.data) +
  geom_histogram(aes(x = gage.height.mean))

```


### Frequency line graph (function: geom_freqpoly)
An alternate to a histogram is a frequency polygon graph (distributions of values for continuous numerical variables). Instead of displaying bars,  counts of continuous variables are displayed as lines. This is advantageous if you want to display multiple variables or categories of variables at once.

```{r, fig.height = 3, fig.width = 4}
# compare multiple values
ggplot(USGS.flow.data) + #dataset plot is based on 
  geom_freqpoly(aes(x = gage.height.mean), bins = 50) + #x: mean stream height; like histogram but continuous (smoothed)
  geom_freqpoly(aes(x = gage.height.min), bins = 50, color = "darkgray") + #compares 2 distributions; specs colour
  geom_freqpoly(aes(x = gage.height.max), bins = 50,  lty = 2) + #specifies line type
  scale_x_continuous(limits = c(0, 10))

#
ggplot(USGS.flow.data) +
  geom_freqpoly(aes(x = gage.height.mean, color = gage.height.mean.approval), bins = 50) + #in aes is whether something is specified based on something in dataframe
  scale_x_continuous(limits = c(0, 10)) +
  theme(legend.position = "top")

```


### Box-and-whisker plots (function: geom_boxplot, geom_violin)
A box-and-whisker plot is yet another alternative to histograms (distributions of values for continuous numerical variables). These plots consist of: 

* A box from the 25th to the 75th percentile of the data, called the interquartile range (IQR).

* A bold line inside the box representing the median value of the data. Whether the median is in the center or off to one side of the IQR will give you an idea about the skewness of your data.

* A line outside of the box representing values falling within 1.5 times the IQR. 

* Points representing outliers, values that fall outside 1.5 times the IQR. 

An alternate option is a violin plot, which displays density distributions, somewhat like a hybrid of the box-and-whiskers and the frequency polygon plot.

```{r, fig.height = 3, fig.width = 4}
# create boxplot of categorical variables
ggplot(USGS.flow.data) +
  geom_boxplot(aes(x = gage.height.mean.approval, y = gage.height.mean))

# create box plot of cts data
ggplot(USGS.flow.data) +
  geom_boxplot(aes(x = gage.height.mean, y = discharge.mean, 
                   group = cut_width(gage.height.mean, 1)))

# create violin plot of categorical data
ggplot(USGS.flow.data) +
  geom_violin(aes(x = gage.height.mean.approval, y = gage.height.mean), 
              draw_quantiles = c(0.25, 0.5, 0.75))
```

### Scatterplot (function: geom_point)
Visualize relationships between continuous numerical variables.

```{r, fig.height = 3, fig.width = 4}
ggplot(USGS.flow.data) +
  geom_point(aes(x = discharge.mean, y = gage.height.mean))

ggplot(USGS.flow.data) +
  geom_point(aes(x = datetime, y = discharge.mean))
```

Question: under what circumstances would it be beneficial to use each of these graph types (bar plot, histogram, frequency polygon, box-and whisker, violin, scatterplot)?

> Answer: 
Box plot: comparing categorical data with a single output for each (mean, sum, etc)
Histogram: frequency distributions of cts data emphasizing values of bins
Frequency polygon: frequency distributions of cts data emphasizing distribution shape
Box and whisker: summarizing data across categorical variables and comparing with emphasis on values
Violin: summarizing data across categorical variables and comparing with emphasis on distribution
Scatterplot: examine relationship between two cts variables (eg. tme series)

## Ending discussion
What did you learn about the USGS discharge dataset today? What separate insights did the different graph types offer?
> Answer: Different graphs offer different types of information about a dataset (distribution, relationships, etc.)


How can multiple options for data exploration inform our understanding of our data?

> Answer: we can look at visualizations of different types of data. They can help us understand how data groups and has patterns that are not apparent in a table. 


Do you see any patterns in the USGS data for the Eno River? What might be responsible for those patterns and/or relationships?

> Answer: Higher mean gage height with higher mean discharge; majority of approved gage height values are between ~0.7 - 3.75;
